ZEN-16125 - Backoff adjusts batch size instead of delay

+++ lib/encoding.go
@@ -39,8 +39,8 @@ func (batch *MetricBatch) MarshalBinary(d *dictionary, doSnappy bool) ([]byte, e
 	buf := new(bytes.Buffer)
 	// Write the API version
 	binary.Write(buf, binary.BigEndian, int8(0))
-	// Write the number of metrics (this could probably be 8-bit)
-	binary.Write(buf, binary.BigEndian, int16(len(batch.Metrics)))
+	// Write the number of metrics
+	binary.Write(buf, binary.BigEndian, int32(len(batch.Metrics)))
 	for _, metric := range batch.Metrics {
 		binary.Write(buf, binary.BigEndian, metric.Timestamp)
 		if metric_name, change = d.get(metric.Metric); change {
